new pages added

+++ StudyNotion/Backend1/controllers/Section.js
@@ -102,33 +102,33 @@ exports.deleteSection = async (req, res) => {
 	try {
 
 		const { sectionId, courseId }  = req.body;
-		// await Course.findByIdAndUpdate(courseId, {
-		// 	$pull: {
-		// 		courseContent: sectionId,
-		// 	}
-		// })
-		// const section = await Section.findById(sectionId);
-		// console.log(sectionId, courseId);
-		// if(!section) {
-		// 	return res.status(404).json({
-		// 		success:false,
-		// 		message:"Section not Found",
-		// 	})
-		// }
+		await Course.findByIdAndUpdate(courseId, {
+			$pull: {
+				courseContent: sectionId,
+			}
+		})
+		const section = await Section.findById(sectionId);
+		console.log(sectionId, courseId);
+		if(!section) {
+			return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+		}
 
 		// //delete sub section
-		// await SubSection.deleteMany({_id: {$in: section.subSection}});
+		await SubSection.deleteMany({_id: {$in: section.subSection}});
 
 		await Section.findByIdAndDelete(sectionId);
 
 		// //find the updated course and return 
-		// const course = await Course.findById(courseId).populate({
-		// 	path:"courseContent",
-		// 	populate: {
-		// 		path: "subSection"
-		// 	}
-		// })
-		// .exec();
+		const course = await Course.findById(courseId).populate({
+			path:"courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+		.exec();
 
 		res.status(200).json({
 			success:true,
